actor/requirements.txt
apify-client>=1.6.0
apify>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.0
